test : GroupService 테스트 작성
