+++ aiku/src/main/java/konkuk/aiku/service/GroupService.java
@@ -73,6 +73,7 @@ public class GroupService {
 
         UserGroup userGroup = checkUserInGroup(user, group);
 
+        
         groupsRepository.deleteUserGroup(userGroup);
         return groupId;
     }
